requirements.txt
alembic==1.8.1
attrs==22.1.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
flasgger==0.9.5
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
jwt==1.3.1
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mistune==2.0.4
packaging==21.3
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.5
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.19.1
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.10.0
